immersionday2021/LAB2_VPCAdv/src/CDK/requirements.txt
aws_cdk.core 
aws_cdk.aws_iam
aws_cdk.aws_ec2
aws_cdk.aws_s3
aws_cdk.aws_s3_assets
